Extract update_release and update_rolling_release into UpdateService
